# backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: maicivy API
  version: 1.0.0
  description: |
    Interactive CV platform with AI-powered cover letter generation.

    This API provides endpoints for:
    - Adaptive CV generation based on themes
    - AI-powered dual cover letter generation (motivation + anti-motivation)
    - Real-time analytics and visitor tracking
    - GitHub integration for automated CV enrichment
    - Professional timeline and profile detection

    ## Rate Limiting
    - AI letter generation: 5 requests/day with 2-minute cooldown between requests
    - Access gate: Requires 3 visits or detected recruiter/tech profile

    ## Authentication
    Session-based authentication using cookies (session_id).
  contact:
    name: Alexi
    email: contact@maicivy.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.maicivy.com
    description: Production server (if deployed)
...
paths:
...
  /api/v1/analytics/realtime:
    get:
      tags:
        - Analytics
      summary: Get real-time statistics
      description: Returns real-time analytics metrics from Redis
      operationId: getRealtimeStats
      responses:
        '200':
          description: Real-time statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    description: Real-time metrics (structure varies)
              example:
                success: true
                data:
                  active_visitors: 5
                  total_page_views: 127
                  cv_views: 42
        '500':
          description: Failed to retrieve stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/analytics/event:
    post:
      tags:
        - Analytics
      summary: Track custom event
      description: Tracks a custom analytics event
      operationId: trackEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - event_type
              properties:
                event_type:
                  type: string
                  example: button_click
                event_data:
                  type: object
                  additionalProperties: true
                  example:
                    button: cta
                    x: 450
                    y: 200
                page_url:
                  type: string
                  example: /cv
      responses:
        '201':
          description: Event tracked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
              example:
                success: true
                message: Event tracked successfully
        '400':
          description: Invalid request or missing visitor session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to track event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ====================
  # GITHUB ENDPOINTS
  # ====================
...
  /api/v1/timeline:
    get:
      tags:
        - Timeline
      summary: Get combined timeline
      description: Returns combined timeline of experiences and projects
      operationId: getTimeline
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
            example: backend
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2020-01-01"
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2025-12-31"
      responses:
        '200':
          description: Timeline events
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      events:
                        type: array
                        items:
                          $ref: '#/components/schemas/TimelineEvent'
                      total:
                        type: integer
                      stats:
                        $ref: '#/components/schemas/TimelineStats'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: string

    # CV
...
    TimelineEvent:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [experience, project]
        title:
          type: string
        subtitle:
          type: string
        content:
          type: string
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
          nullable: true
        tags:
          type: array
          items:
            type: string
        category:
          type: string
        image:
          type: string
          nullable: true
        is_current:
          type: boolean
...
    TimelineStats:
      type: object
      properties:
        total_experiences:
          type: integer
        total_projects:
          type: integer
        categories_breakdown:
          type: object
          additionalProperties:
            type: integer
        years_of_experience:
          type: number
          format: float
        top_technologies:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              count:
                type: integer

    # Profile
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
      description: Session-based authentication using cookies
security:
  - sessionCookie: []